fix merge conflicts + empty skilled groups
