+++ components/SkilledImpact/elements/CampaignContent.js
@@ -1,102 +1,58 @@
 import React from 'react';
-import { View, Text, TouchableOpacity, ActivityIndicator } from 'react-native';
-import styles from '../../../constants/SkilledImpact/CampaignContent';
-import Lightening from '../../../assets/jsicons/bottomnavigation/Lightening';
-import ChevronBottom from '../../../assets/jsicons/miscIcons/ChevronBottom';
+import { View, Text, TouchableOpacity } from 'react-native';
+import styles from '../../../constants/SkilledImpact/OrgSkilledImpactBody';
 import ChevronRight from '../../../assets/jsicons/miscIcons/ChevronRight';
-import CampaignElement from './CampaignElement';
+import { Avatar } from 'react-native-elements';
+import * as moment from 'moment';
 
-import { getCampaignsBySkill } from '../../../store/actions';
-import { connect } from 'react-redux';
+export default React.forwardRef((props, ref) => {
+  const campaign = props.campaign;
 
-class CampaignContent extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      skillExpand: true,
-      isReachable: true,
-      skills: props.skills,
-      campaigns: props.campaigns,
-      loading: false,
-    };
-  }
-
-  componentDidUpdate = async (prevProps) => {
-    if (
-      prevProps.currentUserProfile.skills !==
-      this.props.currentUserProfile.skills
-    ) {
-      this.setState({
-        campaigns: [],
-        loading: true,
-      });
-      const { skills } = this.props.currentUserProfile;
-      if (skills) {
-        for (const skill of skills) {
-          await this.props.getCampaignsBySkill(skill);
-          this.setState({
-            campaigns: this.state.campaigns.concat(this.props.campaignsBySkill),
-          });
-        }
-      }
-      this.setState({
-        loading: false,
-      });
-      console.log('Campaigns Done');
-    }
-  };
-
-  skillExpand = () => {
-    this.setState({ skillExpand: !this.state.skillExpand });
-  };
-
-  render() {
-    const campaignList = this.state.campaigns;
-    return (
-      <View style={styles.itemContainers}>
-        <TouchableOpacity
-          style={styles.itemTitleRow}
-          onPress={this.skillExpand}
-        >
-          <Lightening />
-          <Text style={styles.itemTitleText}>Current Campaigns</Text>
-          <View style={styles.chevronArrowContainer}>
-            {this.state.skillExpand ? <ChevronBottom /> : <ChevronRight />}
-          </View>
-        </TouchableOpacity>
-        {this.state.skillExpand ? (
-          <View style={styles.itemContentBody}>
-            {!this.state.loading ? (
-              campaignList.length !== 0 ? (
-                campaignList.map((campaign, keyIndex) => {
-                  if (campaign) {
-                    return (
-                      <CampaignElement key={keyIndex} campaign={campaign} />
-                    );
-                  }
-                })
-              ) : (
-                <View style={styles.description}>
-                  <Text>
-                    Select your skills above to see available campaigns that
-                    match your skills
-                  </Text>
-                </View>
-              )
-            ) : (
-              <ActivityIndicator style={{ width: '100%', marginBottom: 10 }} />
-            )}
-          </View>
-        ) : null}
+  return (
+    <View style={styles.itemContentRows}>
+      <View style={styles.avatarImageContainer}>
+        <Avatar
+          size={65}
+          rounded
+          source={{
+            uri: campaign.image || undefined,
+          }}
+        />
       </View>
-    );
-  }
-}
-
-const mapStateToProps = (state) => ({
-  currentUserProfile: state.currentUserProfile,
-  campaignsBySkill: state.campaignsBySkill,
+      <View style={styles.campaignRightContainer}>
+        <View style={styles.campaignRow}>
+          {campaign.name ? (
+            <Text style={styles.campaignOrganizationName}>{campaign.name}</Text>
+          ) : (
+            <Text style={styles.campaignOrganizationName}>
+              {campaign.org_name}
+            </Text>
+          )}
+          <View style={styles.chevronArrowContainer}>
+            <ChevronRight />
+          </View>
+        </View>
+        <View style={styles.campaignRow}>
+          {campaign.is_deactivated ? (
+            <View style={styles.closeTag}>
+              <Text style={styles.closeText}>CLOSE</Text>
+            </View>
+          ) : (
+            <View style={styles.openTag}>
+              <Text style={styles.mediumButtonText}>OPEN</Text>
+            </View>
+          )}
+        </View>
+        <View style={styles.campaignRowFooter}>
+          <Text style={styles.intervalPostedText}>
+            {moment
+              .utc(campaign.created_at)
+              .local()
+              .startOf('seconds')
+              .fromNow()}
+          </Text>
+        </View>
+      </View>
+    </View>
+  );
 });
-export default connect(mapStateToProps, {
-  getCampaignsBySkill,
-})(CampaignContent);
